// packages/sif-common-utils/src/dateFormatter.d.ts
export declare const dateFormatter: {
    compact: (date: Date) => string;
    dateShortMonthYear: (date: Date) => string;
    full: (date: Date) => string;
    day: (date: Date) => string;
    dayCompactDate: (date: Date) => string;
    dayDateShortMonthYear: (date: Date) => string;
    dayDateMonthYear: (date: Date) => string;
    dayDateMonth: (date: Date) => string;
    dayDateShortMonth: (date: Date) => string;
};
//# sourceMappingURL=dateFormatter.d.ts.map